requirements.txt
altgraph==0.17.4
cffi==1.17.1
cryptography==43.0.1
distlib==0.3.9
filelock==3.16.1
mysql-connector-python==9.0.0
packaging==24.1
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
setuptools==75.2.0
ttkbootstrap==1.10.1
virtualenv==20.26.6
